requirements.txt
httplib2==0.9.1
requests==2.7.0
python-dateutil==2.5.0